apps/client_backend/pyproject.toml
[tool.poetry]
name = "client-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
python-dotenv = "^1.0.0"
vocode = "0.1.111a2"
elevenlabs = "^0.2.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
